Add Python 3.6 support and drop 3.3
Python 3.3 being near its EOL and not officially
supported by matplotlib since 2.0, the support of
this version in StagPy is dropped. This will allow
for the use of the standard module pathlib (3.4+)
instead of the low level os.path module.

Python 3.6 version of StagPy on the Anaconda cloud
will replace the current Python 3.5 with the release
of Anaconda 4.3.