Use misc.InchoateFiles context manager in plates
This allows for a cleaner handling of the bunch of temporary files
created in plates.py. The corresponding with statement is less clumsy
and more flexible in terms of number of needed temporary files.
Moreover, the context manager ensure temporary files are removed if an
exception is raised (they are, however, kept if an exception is raised
during the final copy of the temporary files to give them their final
names).