New StagpyConfiguration class handling config
An instance of this class at the module level holds configuration
information (default updated with the content of the config file).

It is used to set default for the cmd line arguments parser. The
building of the latter is now much simpler and doesn't require to
parse the arguments twice.