Replace init_config with conf.reset
The function init_config was creating a new configuration object
instead of modifying the existing one. As a result, if stagpy was
imported in several modules before a call to init_config, others
modules were keeping the reference to the old stagpy.conf object
which of course isn't altered by the call to init_config.

In order to keep the state shared through modules, reinitialization
of stagpy.conf should be done with the conf.reset method.