New radial profiles interface
The previous interface was exposing a pandas dataframe, which itself
has an API that's unecessarily complex for our goals and not flexible
as subclassing dataframes properly is difficult.  In particular, this
led to the awkward rprof.get_rprof interface to get RPROF_EXTRA
variables.

This commit fixes these problems by implementing a rprof interface as
member of Step instances which abstracts rprof.get_rprof as well as
misc.get_rbounds.

In particular:

- step.rprofs replaces step.rprof (for consistency with tracers and
  fields attributes which have a plural name);
- items of rprofs are named tuple with values, rad, and meta fields;
- EXTRA variables are seamlessly accessible through the same interface;
- rprof.get_rprof is removed as it is no longer necessary;
- step.rprofs.bounds replaces misc.get_rbounds;
- step.rprofs.centers offers easy access to cell center location;
- step.rprofs.walls offers access to walls location and replaces
  the redges EXTRA variable (which is dropped).