Try reading surrounding h5 files if data not found
File numbers can get slightly out of sync between cores, leading
to erroneous file names in the xdmf file (which is a bug in StagYY).

This commit is a quick fix to be able to read the data anyway by
looking at the previous and next file when a field is not found in
the file where it is expected.

A more robust fix would be to get rid of reading the Data.xmf file
altogether and actually look at where the data is. StagYY output
should be fixed as well.