Fix time series axes boundaries
When requesting an arbitrary tstart and/or tend on time series plots,
the vertical scale was based on the entire series rather than the zoomed
part.  The horizontal scale was also always based on the 't' time series
rather than actual times at which the time series are available (or the
requested tstart/tend).