Scalar fields are plotted at the correct location
Interpolation of scalar field is now turned off by default to avoid
giving the impression that the run is better resolved than it is.
`field.plot_scalar` now exploits walls location by default to draw
scalar fields as they are in FV: a value is assigned to a cell delimited
by those walls. This eliminates the need for an extra layer to close the
spherical annulus. Note that this extra layer is needed when plotting a
smoothed field with the Gouraud shading as the later cannot exploit
walls locations.