New PossiblyCorruptedFileError exception
It is for now only raised when h5py fails with an OSError when
attempting to read a group.  The exception raised by h5py does not
contain information on the file nor the group it failed to read.  This
commit attempts to fix this problem by raising a bespoke exception for
this case.  Note that this hasn't strong guarantees to keep working in
the future as hdf5 doesn't consider error codes to be part of its API,
and h5py maps error codes to Python exceptions.