Cache xdmf files element tree
When processing a large number of snapshots in HDF5, the time spent
parsing the xdmf file completely dominates performance profiles.
This commits is a first step to improve the situation: it caches the
full element tree that results from parsing the xdmf file. This
dramatically speeds up processing, at the cost of a higher memory
footprint. Caching only the information we actually use instead of
the full element tree would reduce the memory footprint.